chore:Folder restructure to be consistent
