+++ src/main/java/com/github/dscpsyl/jgrade2/JGrade2.java
@@ -1,4 +1,4 @@
-package com.github.dscpsyl.jgrade;
+package com.github.dscpsyl.jgrade2;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -8,7 +8,7 @@ import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
-import com.github.dscpsyl.jgrade.gradescope.GradescopeJsonFormatter;
+import com.github.dscpsyl.jgrade2.gradescope.GradescopeJsonFormatter;
 
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
