+++ src/test/java/com/github/dscpsyl/jgrade2/CLITesterExecutionResultTest.java
@@ -1,6 +1,7 @@
-package com.github.dscpsyl.jgrade;
+package com.github.dscpsyl.jgrade2;
 
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class CLITesterExecutionResultTest {
